api-cms 规范接口：根据id查询页面

+++ xcEduService01/xc-service-manage-cms/src/main/java/com/xuecheng/manage_cms/service/PageService.java
@@ -149,7 +149,7 @@ public class PageService {
 
     // 根据id删除页面
     public ResponseResult delete(String id){
-        if (id!=null || !id.equals("")){
+        if (id!=null && !id.equals("")){
             Optional<CmsPage> one = cmsPageRepository.findById(id);
             if (one.isPresent()){
                 cmsPageRepository.deleteById(id);
